Fix ChatBroadcaster constructor parameter order
The constructor was expecting (identity, db, logger, options) but was
being called with (identity, db, options) in the factory, causing the
logger to be undefined. Swapped the parameter order to put options
before logger since options is more commonly used and logger is optional.

This also fixes the multi-user tests which were all failing with
"this.logger.info is not a function" errors. All 6 tests now pass.